# Speed up vector operations