--- a/news/bylengthbenchmark.rst
+++ b/news/bylengthbenchmark.rst
@@ -1,0 +1,3 @@
+**Fixed:**
+
+* speed up many vector operations in particular those involving exact-real elements.