… libflatsurf/src/vector.cc
… libflatsurf/src/vector/algorithm/README.md
… libflatsurf/src/vector/algorithm/base.extension.ipp
… libflatsurf/src/vector/algorithm/base.ipp
… libflatsurf/src/vector/algorithm/exact.extension.ipp
… libflatsurf/src/vector/algorithm/exact.ipp
… libflatsurf/src/vector/algorithm/extension.ipp
… libflatsurf/src/vector/algorithm/with_error.extension.ipp
… libflatsurf/src/vector/algorithm/with_error.ipp
… libflatsurf/src/vector/storage/cartesian.ipp
… libflatsurf/src/vector/storage/forward.ipp
… libflatsurf/src/vector/vector.cc