the original layout of the vector classes was made with the intention of
having lots of different flavours of vectors. In the end we did not do
this but wrote specialized classes such as Chain instead. So vector was
quite overengineered. Now it's much simpler and also a bit faster.

Here, some operations involving exact-real now make better use of Arb
which speeds up some of the predicates quite a bit.

In particular this should make SaddleConnectionsByLength substantially
faster.

```
                                                                        Before  After
VectorUnaryMinus<long long>                                              102 ns   103 ns
VectorUnaryMinus<mpz_class>                                              494 ns   307 ns
VectorUnaryMinus<mpq_class>                                              891 ns   701 ns
VectorUnaryMinus<eantic::renf_elem_class>                                294 ns   224 ns
VectorUnaryMinus<exactreal::Element<exactreal::IntegerRing>>            1671 ns  1094 ns
VectorUnaryMinus<exactreal::Element<exactreal::RationalField>>          2299 ns  1479 ns
VectorUnaryMinus<exactreal::Element<exactreal::NumberField>>            2252 ns  1604 ns
VectorPlus<long long>                                                   98.9 ns  98.6 ns
VectorPlus<mpz_class>                                                    588 ns   591 ns
VectorPlus<mpq_class>                                                   1185 ns  1181 ns
VectorPlus<eantic::renf_elem_class>                                      550 ns   548 ns
VectorPlus<exactreal::Element<exactreal::IntegerRing>>                  1022 ns  1002 ns
VectorPlus<exactreal::Element<exactreal::RationalField>>                1609 ns  1641 ns
VectorPlus<exactreal::Element<exactreal::NumberField>>                  1236 ns  1242 ns
VectorMinus<long long>                                                   201 ns   222 ns
VectorMinus<mpz_class>                                                  1105 ns   902 ns
VectorMinus<mpq_class>                                                  2111 ns  1907 ns
VectorMinus<eantic::renf_elem_class>                                     803 ns   748 ns
VectorMinus<exactreal::Element<exactreal::IntegerRing>>                 2708 ns  2108 ns
VectorMinus<exactreal::Element<exactreal::RationalField>>               3902 ns  3126 ns
VectorMinus<exactreal::Element<exactreal::NumberField>>                 3448 ns  2797 ns
VectorTimesInt<long long>                                               98.5 ns  97.8 ns
VectorTimesInt<mpz_class>                                                585 ns   586 ns
VectorTimesInt<mpq_class>                                               1256 ns  1249 ns
VectorTimesInt<eantic::renf_elem_class>                                  404 ns   404 ns
VectorTimesInt<exactreal::Element<exactreal::IntegerRing>>              1003 ns   991 ns
VectorTimesInt<exactreal::Element<exactreal::RationalField>>            1673 ns  1667 ns
VectorTimesInt<exactreal::Element<exactreal::NumberField>>              1048 ns  1060 ns
VectorTimesMpz<long long>                                                898 ns   889 ns
VectorTimesMpz<mpz_class>                                                589 ns   587 ns
VectorTimesMpz<mpq_class>                                               1664 ns  1652 ns
VectorTimesMpz<eantic::renf_elem_class>                                  463 ns   479 ns
VectorTimesMpz<exactreal::Element<exactreal::IntegerRing>>              1025 ns  1014 ns
VectorTimesMpz<exactreal::Element<exactreal::RationalField>>            2089 ns  2123 ns
VectorTimesMpz<exactreal::Element<exactreal::NumberField>>              1200 ns  1143 ns
VectorDivInt<long long>                                                  115 ns   116 ns
VectorDivInt<mpz_class>                                                  339 ns   338 ns
VectorDivInt<mpq_class>                                                 1258 ns  1253 ns
VectorDivInt<eantic::renf_elem_class>                                    569 ns   562 ns
VectorDivInt<exactreal::Element<exactreal::RationalField>>              1672 ns  1674 ns
VectorDivInt<exactreal::Element<exactreal::NumberField>>                1281 ns  1283 ns
VectorDivMpz<long long>                                                  421 ns   419 ns
VectorDivMpz<mpz_class>                                                  364 ns   371 ns
VectorDivMpz<mpq_class>                                                 1662 ns  1658 ns
VectorDivMpz<eantic::renf_elem_class>                                    632 ns   564 ns
VectorDivMpz<exactreal::Element<exactreal::RationalField>>              2103 ns  2098 ns
VectorDivMpz<exactreal::Element<exactreal::NumberField>>                1371 ns  1291 ns
VectorPerpendicular<long long>                                          98.3 ns  99.1 ns
VectorPerpendicular<mpz_class>                                           393 ns   394 ns
VectorPerpendicular<mpq_class>                                           687 ns   683 ns
VectorPerpendicular<eantic::renf_elem_class>                             243 ns   250 ns
VectorPerpendicular<exactreal::Element<exactreal::IntegerRing>>         1184 ns  1183 ns
VectorPerpendicular<exactreal::Element<exactreal::RationalField>>       1588 ns  1576 ns
VectorPerpendicular<exactreal::Element<exactreal::NumberField>>         1680 ns  1697 ns
VectorArb<long long>                                                     185 ns   148 ns
VectorArb<mpz_class>                                                     171 ns   153 ns
VectorArb<mpq_class>                                                     251 ns   233 ns
VectorArb<eantic::renf_elem_class>                                       897 ns   885 ns
VectorArb<exactreal::Element<exactreal::IntegerRing>>                   1451 ns  1396 ns
VectorArb<exactreal::Element<exactreal::RationalField>>                 1716 ns  1653 ns
VectorArb<exactreal::Element<exactreal::NumberField>>                  20107 ns 19558 ns
VectorCcw<long long>                                                    13.1 ns  12.8 ns
VectorCcw<mpz_class>                                                     601 ns   599 ns
VectorCcw<mpq_class>                                                    1647 ns  1655 ns
VectorCcw<eantic::renf_elem_class>                                      1081 ns  1112 ns
VectorCcw<exactreal::Element<exactreal::IntegerRing>>                   7106 ns  3551 ns
VectorCcw<exactreal::Element<exactreal::RationalField>>                10315 ns  4234 ns
VectorCcw<exactreal::Element<exactreal::NumberField>>                  48421 ns 40881 ns
VectorOrientation<long long>                                            11.8 ns  4.50 ns
VectorOrientation<mpz_class>                                             613 ns   224 ns
VectorOrientation<mpq_class>                                            1723 ns   776 ns
VectorOrientation<eantic::renf_elem_class>                               894 ns   724 ns
VectorOrientation<exactreal::Element<exactreal::IntegerRing>>          10740 ns  3575 ns
VectorOrientation<exactreal::Element<exactreal::RationalField>>        13593 ns  4171 ns
VectorOrientation<exactreal::Element<exactreal::NumberField>>          44097 ns 29423 ns
VectorTimesVector<long long>                                            2.77 ns  2.77 ns
VectorTimesVector<mpz_class>                                             223 ns   220 ns
VectorTimesVector<mpq_class>                                             776 ns   775 ns
VectorTimesVector<eantic::renf_elem_class>                               695 ns   693 ns
VectorTimesVector<exactreal::Element<exactreal::IntegerRing>>           4557 ns  4528 ns
VectorTimesVector<exactreal::Element<exactreal::RationalField>>         7234 ns  7218 ns
VectorTimesVector<exactreal::Element<exactreal::NumberField>>          31337 ns 33976 ns
VectorLtBound<long long>                                                 409 ns   405 ns
VectorLtBound<mpz_class>                                                 923 ns   918 ns
VectorLtBound<mpq_class>                                                2029 ns  2053 ns
VectorLtBound<eantic::renf_elem_class>                                  1317 ns  1293 ns
VectorLtBound<exactreal::Element<exactreal::IntegerRing>>              10099 ns  3184 ns
VectorLtBound<exactreal::Element<exactreal::RationalField>>            13131 ns  3551 ns
VectorLtBound<exactreal::Element<exactreal::NumberField>>              46298 ns 16566 ns
VectorEqVector<long long>                                               12.9 ns  14.9 ns
VectorEqVector<mpz_class>                                                399 ns   400 ns
VectorEqVector<mpq_class>                                                795 ns   799 ns
VectorEqVector<eantic::renf_elem_class>                                  220 ns   219 ns
VectorEqVector<exactreal::Element<exactreal::IntegerRing>>              1222 ns  1214 ns
VectorEqVector<exactreal::Element<exactreal::RationalField>>            1638 ns  1629 ns
VectorEqVector<exactreal::Element<exactreal::NumberField>>              1463 ns  1461 ns
```